More explicit constants definition
- Rename the file (tool -> position)
